1.首先修改了登录Login的登录条件，多了一个判断步骤 2.在界面UI上，添加了BuyerInterface（采购员界面）、ReplenishStock（进货界面）、SaleReturn（退货界面） 3.其中采购员界面，包括查询、进货、退货、刷新四个按钮，暂时实现了查询、进货和刷新三个按钮的功能。其中查询可以根据输入框里面的商品名称查询到该商品的信息，进货则是可以对商品名称（如果没有）、商品库存、进货基本表和进货细节表进行修改，而刷新则是重新执行SQL语句查询数据库中的商品表的信息
以下为关于进货按钮实现的说明：

创建的文件和文件夹：
- 创建了manager层，包含一个名为ReplenishManager（进货的信息封装）的类文件，其作用是对service层要用到的信息进行封装
- 创建了service层，包含一个ReplenishStockService接口和它的实现类。作用是进行事务的处理
- 创建了dao层，包含四个dao接口和四个dao接口的实现类，作用是实现JDBC，包括对shop数据库的商品表的添加、商品库存的修改、进货基本表和进货细节表的信息增加

以下为具体步骤：
在监听到“进货”按钮被点击后，通过判断进货的商品是否已经存在于商品表而创建不同的ReplenishManager的对象，之后在manager层对需要的数据进行封装并传到service层，service主要是进行事务的开启和结束，而dao层则是执行JDBC，对数据库进行修改

+++ src/main/java/cn/edu/guet/WeShop/bean/Item.java
@@ -5,7 +5,7 @@ import java.util.Objects;
 public class Item {
     private String id;
     private String title;
-    private Float price;
+    private double price;
     private String description;
     private int sales;
 
@@ -36,11 +36,11 @@ public class Item {
         this.title = title;
     }
 
-    public Float getPrice() {
+    public double getPrice() {
         return price;
     }
 
-    public void setPrice(Float price) {
+    public void setPrice(double price) {
         this.price = price;
     }
